Reorganized 'layers.py'
