+++ mnist.py
@@ -2,8 +2,9 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-
-from layers import *
+import numpy as np
+import tensorflow as tf
+from graph.graph import Graph
 from tensorflow_utils import *
 from testing import *
